fix: moved homepage components to folder
