+++ src/components/home/Slant.js
